Currency Note Recognition/Predictions/predictions-backup/requirements.txt
absl-py==1.0.0
aiohttp==3.6.3
alabaster==0.7.12
anaconda-client==1.2.2
appdirs==1.4.4
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.2.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.6.0
cryptography==2.7
cycler==0.11.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
Flask-SQLAlchemy==2.5.1
fonttools==4.28.1
fsspec==0.5.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.7
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
googletrans==3.0.0
greenlet==0.4.15
grpcio==1.27.2
h11==0.9.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.12.22
html5lib==1.0.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
imageio==2.11.0
imagesize==1.1.0
importlib-metadata==2.0.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.3.2
lazy-object-proxy==1.4.2
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.5.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.6
multipledispatch==0.6.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.6.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.21.4
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.0
opt-einsum==3.1.0
packaging==21.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.13.0
psutil==5.6.3
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
tensorflow==2.2.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.1.0
tensorflow-metadata==0.14.0
git+git://github.com/qubvel/segmentation_models@master
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tifffile==2021.11.2
tomli==1.2.2
toolz==0.10.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.24.2
virtualenv==20.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
